+++ app/src/main/java/com/ebata_shota/holdemstacktracker/domain/model/PlayerState.kt
@@ -3,6 +3,6 @@ package com.ebata_shota.holdemstacktracker.domain.model
 data class PlayerState(
     val id: PlayerId,
     val name: String,
-    val stack: Float,
+    val stack: Double,
     val isLeaved: Boolean
 )
