+++ app/src/main/java/com/ebata_shota/holdemstacktracker/domain/model/RuleState.kt
@@ -4,8 +4,8 @@ sealed interface RuleState {
     val betViewMode: BetViewMode
 
     data class LingGame(
-        val sbSize: Float,
-        val bbSize: Float,
+        val sbSize: Double,
+        val bbSize: Double,
         override val betViewMode: BetViewMode
     ) : RuleState
 }
