+++ app/src/main/java/com/ebata_shota/holdemstacktracker/domain/usecase/GetPodStateListUseCase.kt
@@ -11,6 +11,6 @@ interface GetPodStateListUseCase {
      */
     fun invoke(
         podStateList: List<PodState>,
-        pendingBetPerPlayer: Map<PlayerId, Float>
+        pendingBetPerPlayer: Map<PlayerId, Double>
     ): List<PodState>
 }
