+++ app/src/main/java/com/ebata_shota/holdemstacktracker/domain/usecase/impl/GetPendingBetPerPlayerUseCaseImpl.kt
@@ -11,12 +11,15 @@ class GetPendingBetPerPlayerUseCaseImpl
 constructor(
     private val getMaxBetSize: GetMaxBetSizeUseCase
 ) : GetPendingBetPerPlayerUseCase {
-    override fun invoke(playerOrder: List<PlayerId>, actionStateList: List<BetPhaseActionState>): Map<PlayerId, Float> {
+    override fun invoke(
+        playerOrder: List<PlayerId>,
+        actionStateList: List<BetPhaseActionState>
+    ): Map<PlayerId, Double> {
         return playerOrder.associateWith { playerId ->
             // このフェーズでの、特定プレイヤーのアクション一覧を取得
             val playerActionList = actionStateList.filter { it.playerId == playerId }
             // プレイヤーの最大ベットサイズ = そのプレイヤーの最終的なベットサイズ
             getMaxBetSize.invoke(playerActionList)
-        }.filter { it.value > 0.0f }
+        }.filter { it.value > 0.0 }
     }
 }
