+++ app/src/test/java/com/ebata_shota/holdemstacktracker/domain/usecase/GetNextPlayerStateListUseCaseImplTest.kt
@@ -34,7 +34,7 @@ class GetNextPlayerStateListUseCaseImplTest {
     }
 
     private fun executeAndAssert(
-        pendingBetPerPlayer: Map<PlayerId, Float> = emptyMap(),
+        pendingBetPerPlayer: Map<PlayerId, Double> = emptyMap(),
         players: List<PlayerState>,
         action: BetPhaseActionState,
         expected: List<PlayerState>,
@@ -56,18 +56,18 @@ class GetNextPlayerStateListUseCaseImplTest {
         // prepare
         every { prefRepository.myPlayerId } returns flowOf("0")
         val players = listOf(
-            PlayerState(id = PlayerId("0"), name = "0", stack = 1000.0f, isLeaved = false),
-            PlayerState(id = PlayerId("1"), name = "1", stack = 1000.0f, isLeaved = false),
-            PlayerState(id = PlayerId("2"), name = "2", stack = 1000.0f, isLeaved = false),
+            PlayerState(id = PlayerId("0"), name = "0", stack = 1000.0, isLeaved = false),
+            PlayerState(id = PlayerId("1"), name = "1", stack = 1000.0, isLeaved = false),
+            PlayerState(id = PlayerId("2"), name = "2", stack = 1000.0, isLeaved = false),
         )
-        val action = BetPhaseActionState.Blind(actionId = 0L, playerId = PlayerId("0"), betSize = 100.0f)
+        val action = BetPhaseActionState.Blind(actionId = 0L, playerId = PlayerId("0"), betSize = 100.0)
         executeAndAssert(
             players = players,
             action = action,
             expected = listOf(
-                PlayerState(id = PlayerId("0"), name = "0", stack = 900.0f, isLeaved = false),
-                PlayerState(id = PlayerId("1"), name = "1", stack = 1000.0f, isLeaved = false),
-                PlayerState(id = PlayerId("2"), name = "2", stack = 1000.0f, isLeaved = false),
+                PlayerState(id = PlayerId("0"), name = "0", stack = 900.0, isLeaved = false),
+                PlayerState(id = PlayerId("1"), name = "1", stack = 1000.0, isLeaved = false),
+                PlayerState(id = PlayerId("2"), name = "2", stack = 1000.0, isLeaved = false),
             )
         )
     }
@@ -77,18 +77,18 @@ class GetNextPlayerStateListUseCaseImplTest {
         // prepare
         every { prefRepository.myPlayerId } returns flowOf("0")
         val players = listOf(
-            PlayerState(id = PlayerId("0"), name = "0", stack = 1000.0f, isLeaved = false),
-            PlayerState(id = PlayerId("1"), name = "1", stack = 1000.0f, isLeaved = false),
-            PlayerState(id = PlayerId("2"), name = "2", stack = 1000.0f, isLeaved = false),
+            PlayerState(id = PlayerId("0"), name = "0", stack = 1000.0, isLeaved = false),
+            PlayerState(id = PlayerId("1"), name = "1", stack = 1000.0, isLeaved = false),
+            PlayerState(id = PlayerId("2"), name = "2", stack = 1000.0, isLeaved = false),
         )
-        val action = BetPhaseActionState.Call(actionId = 0L, playerId = PlayerId("0"), betSize = 100.0f)
+        val action = BetPhaseActionState.Call(actionId = 0L, playerId = PlayerId("0"), betSize = 100.0)
         executeAndAssert(
             players = players,
             action = action,
             expected = listOf(
-                PlayerState(id = PlayerId("0"), name = "0", stack = 900.0f, isLeaved = false),
-                PlayerState(id = PlayerId("1"), name = "1", stack = 1000.0f, isLeaved = false),
-                PlayerState(id = PlayerId("2"), name = "2", stack = 1000.0f, isLeaved = false),
+                PlayerState(id = PlayerId("0"), name = "0", stack = 900.0, isLeaved = false),
+                PlayerState(id = PlayerId("1"), name = "1", stack = 1000.0, isLeaved = false),
+                PlayerState(id = PlayerId("2"), name = "2", stack = 1000.0, isLeaved = false),
             )
         )
     }
@@ -98,18 +98,18 @@ class GetNextPlayerStateListUseCaseImplTest {
         // prepare
         every { prefRepository.myPlayerId } returns flowOf("0")
         val players = listOf(
-            PlayerState(id = PlayerId("0"), name = "0", stack = 1000.0f, isLeaved = false),
-            PlayerState(id = PlayerId("1"), name = "1", stack = 1000.0f, isLeaved = false),
-            PlayerState(id = PlayerId("2"), name = "2", stack = 1000.0f, isLeaved = false),
+            PlayerState(id = PlayerId("0"), name = "0", stack = 1000.0, isLeaved = false),
+            PlayerState(id = PlayerId("1"), name = "1", stack = 1000.0, isLeaved = false),
+            PlayerState(id = PlayerId("2"), name = "2", stack = 1000.0, isLeaved = false),
         )
-        val action = BetPhaseActionState.Raise(actionId = 0L, playerId = PlayerId("0"), betSize = 400.0f)
+        val action = BetPhaseActionState.Raise(actionId = 0L, playerId = PlayerId("0"), betSize = 400.0)
         executeAndAssert(
             players = players,
             action = action,
             expected = listOf(
-                PlayerState(id = PlayerId("0"), name = "0", stack = 600.0f, isLeaved = false),
-                PlayerState(id = PlayerId("1"), name = "1", stack = 1000.0f, isLeaved = false),
-                PlayerState(id = PlayerId("2"), name = "2", stack = 1000.0f, isLeaved = false),
+                PlayerState(id = PlayerId("0"), name = "0", stack = 600.0, isLeaved = false),
+                PlayerState(id = PlayerId("1"), name = "1", stack = 1000.0, isLeaved = false),
+                PlayerState(id = PlayerId("2"), name = "2", stack = 1000.0, isLeaved = false),
             )
         )
     }
@@ -119,18 +119,18 @@ class GetNextPlayerStateListUseCaseImplTest {
         // prepare
         every { prefRepository.myPlayerId } returns flowOf("0")
         val players = listOf(
-            PlayerState(id = PlayerId("0"), name = "0", stack = 1000.0f, isLeaved = false),
-            PlayerState(id = PlayerId("1"), name = "1", stack = 1000.0f, isLeaved = false),
-            PlayerState(id = PlayerId("2"), name = "2", stack = 1000.0f, isLeaved = false),
+            PlayerState(id = PlayerId("0"), name = "0", stack = 1000.0, isLeaved = false),
+            PlayerState(id = PlayerId("1"), name = "1", stack = 1000.0, isLeaved = false),
+            PlayerState(id = PlayerId("2"), name = "2", stack = 1000.0, isLeaved = false),
         )
-        val action = BetPhaseActionState.Bet(actionId = 0L, playerId = PlayerId("0"), betSize = 100.0f)
+        val action = BetPhaseActionState.Bet(actionId = 0L, playerId = PlayerId("0"), betSize = 100.0)
         executeAndAssert(
             players = players,
             action = action,
             expected = listOf(
-                PlayerState(id = PlayerId("0"), name = "0", stack = 900.0f, isLeaved = false),
-                PlayerState(id = PlayerId("1"), name = "1", stack = 1000.0f, isLeaved = false),
-                PlayerState(id = PlayerId("2"), name = "2", stack = 1000.0f, isLeaved = false),
+                PlayerState(id = PlayerId("0"), name = "0", stack = 900.0, isLeaved = false),
+                PlayerState(id = PlayerId("1"), name = "1", stack = 1000.0, isLeaved = false),
+                PlayerState(id = PlayerId("2"), name = "2", stack = 1000.0, isLeaved = false),
             )
         )
     }
@@ -140,18 +140,18 @@ class GetNextPlayerStateListUseCaseImplTest {
         // prepare
         every { prefRepository.myPlayerId } returns flowOf("0")
         val players = listOf(
-            PlayerState(id = PlayerId("0"), name = "0", stack = 1000.0f, isLeaved = false),
-            PlayerState(id = PlayerId("1"), name = "1", stack = 1000.0f, isLeaved = false),
-            PlayerState(id = PlayerId("2"), name = "2", stack = 1000.0f, isLeaved = false),
+            PlayerState(id = PlayerId("0"), name = "0", stack = 1000.0, isLeaved = false),
+            PlayerState(id = PlayerId("1"), name = "1", stack = 1000.0, isLeaved = false),
+            PlayerState(id = PlayerId("2"), name = "2", stack = 1000.0, isLeaved = false),
         )
-        val action = BetPhaseActionState.AllIn(actionId = 0L, playerId = PlayerId("0"), betSize = 1000.0f)
+        val action = BetPhaseActionState.AllIn(actionId = 0L, playerId = PlayerId("0"), betSize = 1000.0)
         executeAndAssert(
             players = players,
             action = action,
             expected = listOf(
-                PlayerState(id = PlayerId("0"), name = "0", stack = 0.0f, isLeaved = false),
-                PlayerState(id = PlayerId("1"), name = "1", stack = 1000.0f, isLeaved = false),
-                PlayerState(id = PlayerId("2"), name = "2", stack = 1000.0f, isLeaved = false),
+                PlayerState(id = PlayerId("0"), name = "0", stack = 0.0, isLeaved = false),
+                PlayerState(id = PlayerId("1"), name = "1", stack = 1000.0, isLeaved = false),
+                PlayerState(id = PlayerId("2"), name = "2", stack = 1000.0, isLeaved = false),
             )
         )
     }
@@ -161,18 +161,18 @@ class GetNextPlayerStateListUseCaseImplTest {
         // prepare
         every { prefRepository.myPlayerId } returns flowOf("0")
         val players = listOf(
-            PlayerState(id = PlayerId("0"), name = "0", stack = 1000.0f, isLeaved = false),
-            PlayerState(id = PlayerId("1"), name = "1", stack = 1000.0f, isLeaved = false),
-            PlayerState(id = PlayerId("2"), name = "2", stack = 1000.0f, isLeaved = false),
+            PlayerState(id = PlayerId("0"), name = "0", stack = 1000.0, isLeaved = false),
+            PlayerState(id = PlayerId("1"), name = "1", stack = 1000.0, isLeaved = false),
+            PlayerState(id = PlayerId("2"), name = "2", stack = 1000.0, isLeaved = false),
         )
         val action = BetPhaseActionState.Check(actionId = 0L, playerId = PlayerId("0"))
         executeAndAssert(
             players = players,
             action = action,
             expected = listOf(
-                PlayerState(id = PlayerId("0"), name = "0", stack = 1000.0f, isLeaved = false),
-                PlayerState(id = PlayerId("1"), name = "1", stack = 1000.0f, isLeaved = false),
-                PlayerState(id = PlayerId("2"), name = "2", stack = 1000.0f, isLeaved = false),
+                PlayerState(id = PlayerId("0"), name = "0", stack = 1000.0, isLeaved = false),
+                PlayerState(id = PlayerId("1"), name = "1", stack = 1000.0, isLeaved = false),
+                PlayerState(id = PlayerId("2"), name = "2", stack = 1000.0, isLeaved = false),
             )
         )
     }
@@ -182,18 +182,18 @@ class GetNextPlayerStateListUseCaseImplTest {
         // prepare
         every { prefRepository.myPlayerId } returns flowOf("0")
         val players = listOf(
-            PlayerState(id = PlayerId("0"), name = "0", stack = 1000.0f, isLeaved = false),
-            PlayerState(id = PlayerId("1"), name = "1", stack = 1000.0f, isLeaved = false),
-            PlayerState(id = PlayerId("2"), name = "2", stack = 1000.0f, isLeaved = false),
+            PlayerState(id = PlayerId("0"), name = "0", stack = 1000.0, isLeaved = false),
+            PlayerState(id = PlayerId("1"), name = "1", stack = 1000.0, isLeaved = false),
+            PlayerState(id = PlayerId("2"), name = "2", stack = 1000.0, isLeaved = false),
         )
         val action = BetPhaseActionState.Fold(actionId = 0L, playerId = PlayerId("0"))
         executeAndAssert(
             players = players,
             action = action,
             expected = listOf(
-                PlayerState(id = PlayerId("0"), name = "0", stack = 1000.0f, isLeaved = false),
-                PlayerState(id = PlayerId("1"), name = "1", stack = 1000.0f, isLeaved = false),
-                PlayerState(id = PlayerId("2"), name = "2", stack = 1000.0f, isLeaved = false),
+                PlayerState(id = PlayerId("0"), name = "0", stack = 1000.0, isLeaved = false),
+                PlayerState(id = PlayerId("1"), name = "1", stack = 1000.0, isLeaved = false),
+                PlayerState(id = PlayerId("2"), name = "2", stack = 1000.0, isLeaved = false),
             )
         )
     }
@@ -203,18 +203,18 @@ class GetNextPlayerStateListUseCaseImplTest {
         // prepare
         every { prefRepository.myPlayerId } returns flowOf("0")
         val players = listOf(
-            PlayerState(id = PlayerId("0"), name = "0", stack = 1000.0f, isLeaved = false),
-            PlayerState(id = PlayerId("1"), name = "1", stack = 1000.0f, isLeaved = false),
-            PlayerState(id = PlayerId("2"), name = "2", stack = 1000.0f, isLeaved = false),
+            PlayerState(id = PlayerId("0"), name = "0", stack = 1000.0, isLeaved = false),
+            PlayerState(id = PlayerId("1"), name = "1", stack = 1000.0, isLeaved = false),
+            PlayerState(id = PlayerId("2"), name = "2", stack = 1000.0, isLeaved = false),
         )
         val action = BetPhaseActionState.FoldSkip(actionId = 0L, playerId = PlayerId("0"))
         executeAndAssert(
             players = players,
             action = action,
             expected = listOf(
-                PlayerState(id = PlayerId("0"), name = "0", stack = 1000.0f, isLeaved = false),
-                PlayerState(id = PlayerId("1"), name = "1", stack = 1000.0f, isLeaved = false),
-                PlayerState(id = PlayerId("2"), name = "2", stack = 1000.0f, isLeaved = false),
+                PlayerState(id = PlayerId("0"), name = "0", stack = 1000.0, isLeaved = false),
+                PlayerState(id = PlayerId("1"), name = "1", stack = 1000.0, isLeaved = false),
+                PlayerState(id = PlayerId("2"), name = "2", stack = 1000.0, isLeaved = false),
             )
         )
     }
@@ -224,24 +224,24 @@ class GetNextPlayerStateListUseCaseImplTest {
         // prepare
         every { prefRepository.myPlayerId } returns flowOf("0")
         val pendingBetPerPlayer = mapOf(
-            PlayerId("0") to 100.0f,
-            PlayerId("1") to 200.0f,
-            PlayerId("2") to 200.0f,
+            PlayerId("0") to 100.0,
+            PlayerId("1") to 200.0,
+            PlayerId("2") to 200.0,
         )
         val players = listOf(
-            PlayerState(id = PlayerId("0"), name = "0", stack = 900.0f, isLeaved = false),
-            PlayerState(id = PlayerId("1"), name = "1", stack = 800.0f, isLeaved = false),
-            PlayerState(id = PlayerId("2"), name = "2", stack = 800.0f, isLeaved = false),
+            PlayerState(id = PlayerId("0"), name = "0", stack = 900.0, isLeaved = false),
+            PlayerState(id = PlayerId("1"), name = "1", stack = 800.0, isLeaved = false),
+            PlayerState(id = PlayerId("2"), name = "2", stack = 800.0, isLeaved = false),
         )
-        val action = BetPhaseActionState.Call(actionId = 0L, playerId = PlayerId("0"), betSize = 200.0f)
+        val action = BetPhaseActionState.Call(actionId = 0L, playerId = PlayerId("0"), betSize = 200.0)
         executeAndAssert(
             pendingBetPerPlayer = pendingBetPerPlayer,
             players = players,
             action = action,
             expected = listOf(
-                PlayerState(id = PlayerId("0"), name = "0", stack = 800.0f, isLeaved = false),
-                PlayerState(id = PlayerId("1"), name = "1", stack = 800.0f, isLeaved = false),
-                PlayerState(id = PlayerId("2"), name = "2", stack = 800.0f, isLeaved = false),
+                PlayerState(id = PlayerId("0"), name = "0", stack = 800.0, isLeaved = false),
+                PlayerState(id = PlayerId("1"), name = "1", stack = 800.0, isLeaved = false),
+                PlayerState(id = PlayerId("2"), name = "2", stack = 800.0, isLeaved = false),
             )
         )
     }
@@ -251,24 +251,24 @@ class GetNextPlayerStateListUseCaseImplTest {
         // prepare
         every { prefRepository.myPlayerId } returns flowOf("0")
         val pendingBetPerPlayer = mapOf(
-            PlayerId("0") to 100.0f,
-            PlayerId("1") to 200.0f,
-            PlayerId("2") to 200.0f,
+            PlayerId("0") to 100.0,
+            PlayerId("1") to 200.0,
+            PlayerId("2") to 200.0,
         )
         val players = listOf(
-            PlayerState(id = PlayerId("0"), name = "0", stack = 900.0f, isLeaved = false),
-            PlayerState(id = PlayerId("1"), name = "1", stack = 800.0f, isLeaved = false),
-            PlayerState(id = PlayerId("2"), name = "2", stack = 800.0f, isLeaved = false),
+            PlayerState(id = PlayerId("0"), name = "0", stack = 900.0, isLeaved = false),
+            PlayerState(id = PlayerId("1"), name = "1", stack = 800.0, isLeaved = false),
+            PlayerState(id = PlayerId("2"), name = "2", stack = 800.0, isLeaved = false),
         )
-        val action = BetPhaseActionState.AllIn(actionId = 0L, playerId = PlayerId("0"), betSize = 1000.0f)
+        val action = BetPhaseActionState.AllIn(actionId = 0L, playerId = PlayerId("0"), betSize = 1000.0)
         executeAndAssert(
             pendingBetPerPlayer = pendingBetPerPlayer,
             players = players,
             action = action,
             expected = listOf(
-                PlayerState(id = PlayerId("0"), name = "0", stack = 0.0f, isLeaved = false),
-                PlayerState(id = PlayerId("1"), name = "1", stack = 800.0f, isLeaved = false),
-                PlayerState(id = PlayerId("2"), name = "2", stack = 800.0f, isLeaved = false),
+                PlayerState(id = PlayerId("0"), name = "0", stack = 0.0, isLeaved = false),
+                PlayerState(id = PlayerId("1"), name = "1", stack = 800.0, isLeaved = false),
+                PlayerState(id = PlayerId("2"), name = "2", stack = 800.0, isLeaved = false),
             )
         )
     }
@@ -277,21 +277,21 @@ class GetNextPlayerStateListUseCaseImplTest {
     fun action_SB_Check_and_BB_Bet() {
         // prepare
         every { prefRepository.myPlayerId } returns flowOf("1")
-        val pendingBetPerPlayer = emptyMap<PlayerId, Float>()
+        val pendingBetPerPlayer = emptyMap<PlayerId, Double>()
         val players = listOf(
-            PlayerState(id = PlayerId("0"), name = "0", stack = 800.0f, isLeaved = false),
-            PlayerState(id = PlayerId("1"), name = "1", stack = 800.0f, isLeaved = false),
-            PlayerState(id = PlayerId("2"), name = "2", stack = 800.0f, isLeaved = false),
+            PlayerState(id = PlayerId("0"), name = "0", stack = 800.0, isLeaved = false),
+            PlayerState(id = PlayerId("1"), name = "1", stack = 800.0, isLeaved = false),
+            PlayerState(id = PlayerId("2"), name = "2", stack = 800.0, isLeaved = false),
         )
-        val action = BetPhaseActionState.Bet(actionId = 0L, playerId = PlayerId("1"), betSize = 200.0f)
+        val action = BetPhaseActionState.Bet(actionId = 0L, playerId = PlayerId("1"), betSize = 200.0)
         executeAndAssert(
             pendingBetPerPlayer = pendingBetPerPlayer,
             players = players,
             action = action,
             expected = listOf(
-                PlayerState(id = PlayerId("0"), name = "0", stack = 800.0f, isLeaved = false),
-                PlayerState(id = PlayerId("1"), name = "1", stack = 600.0f, isLeaved = false),
-                PlayerState(id = PlayerId("2"), name = "2", stack = 800.0f, isLeaved = false),
+                PlayerState(id = PlayerId("0"), name = "0", stack = 800.0, isLeaved = false),
+                PlayerState(id = PlayerId("1"), name = "1", stack = 600.0, isLeaved = false),
+                PlayerState(id = PlayerId("2"), name = "2", stack = 800.0, isLeaved = false),
             )
         )
     }
